JBR-10050 Skip signing of symlinks

+++ jb/project/tools/mac/scripts/sign.sh
@@ -185,15 +185,19 @@ if [ "$JB_SIGN" = true ]; then for f in \
   if [ -d "$APPLICATION_PATH/$f" ]; then
     find "$APPLICATION_PATH/$f" \( -name '*.framework' -o -name '*.app' \) -maxdepth 1 | while read -r line
       do
-        log "Signing '$line':"
-        tar -pczf tmp-to-sign.tar.gz -C "$(dirname "$line")" "$(basename "$line")"
-        "$SIGN_UTILITY" --timestamp \
-            -v -s "$JB_DEVELOPER_CERT" --options=runtime \
-            --force \
-            --entitlements "$SCRIPT_DIR/entitlements_jcef.xml" tmp-to-sign.tar.gz || exit 1
-        rm -rf "$line"
-        tar -xzf tmp-to-sign.tar.gz --directory "$(dirname "$line")"
-        rm -f tmp-to-sign.tar.gz
+        if [ -L "$line" ]; then
+          log "Skip symlink signing: $line"
+        else
+          log "Signing '$line':"
+          tar -pczf tmp-to-sign.tar.gz -C "$(dirname "$line")" "$(basename "$line")"
+          "$SIGN_UTILITY" --timestamp \
+              -v -s "$JB_DEVELOPER_CERT" --options=runtime \
+              --force \
+              --entitlements "$SCRIPT_DIR/entitlements_jcef.xml" tmp-to-sign.tar.gz || exit 1
+          rm -rf "$line"
+          tar -xzf tmp-to-sign.tar.gz --directory "$(dirname "$line")"
+          rm -f tmp-to-sign.tar.gz
+        fi
       done
   fi
 done; fi
